+++ libraries/common/src/main/java/androidx/media3/common/PlaybackException.java
@@ -220,6 +220,10 @@ public class PlaybackException extends Exception implements Bundleable {
   /** Caused by an AudioTrack write operation failure. */
   public static final int ERROR_CODE_AUDIO_TRACK_WRITE_FAILED = 5002;
 
+  // TODO(b/299907254): Stabilize error code, remove @UnstableApi annotation, and add to IntDef
+  /** Caused by an AudioTrack write operation failure in offload mode. */
+  @UnstableApi public static final int ERROR_CODE_AUDIO_TRACK_OFFLOAD_WRITE_FAILED = 5003;
+
   // DRM errors (6xxx).
 
   /** Caused by an unspecified error related to DRM protection. */
@@ -324,6 +328,8 @@ public class PlaybackException extends Exception implements Bundleable {
         return "ERROR_CODE_AUDIO_TRACK_INIT_FAILED";
       case ERROR_CODE_AUDIO_TRACK_WRITE_FAILED:
         return "ERROR_CODE_AUDIO_TRACK_WRITE_FAILED";
+      case ERROR_CODE_AUDIO_TRACK_OFFLOAD_WRITE_FAILED:
+        return "ERROR_CODE_AUDIO_TRACK_OFFLOAD_WRITE_FAILED";
       case ERROR_CODE_DRM_UNSPECIFIED:
         return "ERROR_CODE_DRM_UNSPECIFIED";
       case ERROR_CODE_DRM_SCHEME_UNSUPPORTED: