+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/DefaultAudioSink.java
@@ -1163,9 +1163,17 @@ public final class DefaultAudioSink implements AudioSink {
       int error = bytesWrittenOrError;
 
       // Treat a write error on a previously successful offload channel as recoverable
-      // without disabling offload. Offload will be disabled when a new AudioTrack is created,
-      // if no longer supported.
-      boolean isRecoverable = isAudioTrackDeadObject(error) && getWrittenFrames() > 0;
+      // without disabling offload. Offload will be disabled if offload channel was not successfully
+      // written to or when a new AudioTrack is created, if no longer supported.
+      boolean isRecoverable = false;
+      if (isAudioTrackDeadObject(error)) {
+        if (getWrittenFrames() > 0) {
+          isRecoverable = true;
+        } else if (isOffloadedPlayback(audioTrack)) {
+          maybeDisableOffload();
+          isRecoverable = true;
+        }
+      }
 
       WriteException e = new WriteException(error, configuration.inputFormat, isRecoverable);
       if (listener != null) {